+++ API/Configuration/MappingProfile.cs
@@ -21,13 +21,11 @@ using Models;
 
 namespace API.Configuration
 {
-
     /// <summary>
     ///     This profiles adds every resource mapping.
     /// </summary>
     public class MappingProfile : Profile
     {
-
         /// <summary>
         ///     Create a map for every resource mapping.
         /// </summary>
@@ -48,8 +46,15 @@ namespace API.Configuration
 
             CreateMap<HighlightResource, Highlight>();
             CreateMap<Highlight, HighlightResourceResult>();
+
+            CreateMap<RoleResource, Role>();
+            CreateMap<Role, RoleResourceResult>();
+
+            CreateMap<EmbeddedProjectResource, EmbeddedProject>();
+            CreateMap<EmbeddedProject, EmbeddedProjectResourceResult>();
+
+            CreateMap<RoleScopeResource, RoleScope>();
+            CreateMap<RoleScope, RoleScopeResource>();
         }
-
     }
-
 }