+++ Data/ApplicationDbContext.cs
@@ -15,6 +15,7 @@
 * If not, see https://www.gnu.org/licenses/lgpl-3.0.txt
 */
 
+using System.Collections.Generic;
 using Data.Configurations;
 using Data.Helpers;
 using Microsoft.EntityFrameworkCore;
@@ -22,19 +23,58 @@ using Models;
 
 namespace Data
 {
-
+    /// <summary>
+    /// ApplicationDatabaseContext
+    /// </summary>
+    /// <seealso cref="Microsoft.EntityFrameworkCore.DbContext" />
     public class ApplicationDbContext : DbContext
     {
 
         public ApplicationDbContext(DbContextOptions<ApplicationDbContext> options) : base(options) { }
 
+        /// <summary>
+        /// Gets or sets the user.
+        /// </summary>
+        /// <value>
+        /// The user.
+        /// </value>
         public DbSet<User> User { get; set; }
 
+        /// <summary>
+        /// Gets or sets the project.
+        /// </summary>
+        /// <value>
+        /// The project.
+        /// </value>
         public DbSet<Project> Project { get; set; }
-
+        /// <summary>
+        /// Gets or sets the collaborators.
+        /// </summary>
+        /// <value>
+        /// The collaborators.
+        /// </value>
         public DbSet<Collaborator> Collaborators { get; set; }
-
+        /// <summary>
+        /// Gets or sets the highlight.
+        /// </summary>
+        /// <value>
+        /// The highlight.
+        /// </value>
         public DbSet<Highlight> Highlight { get; set; }
+        /// <summary>
+        /// Gets or sets the embedded project.
+        /// </summary>
+        /// <value>
+        /// The embedded project.
+        /// </value>
+        public DbSet<EmbeddedProject> EmbeddedProject { get; set; }
+        /// <summary>
+        /// Gets or sets the role.
+        /// </summary>
+        /// <value>
+        /// The role.
+        /// </value>
+        public DbSet<Role> Role { get; set; }
 
         protected override void OnModelCreating(ModelBuilder modelBuilder)
         {
@@ -43,9 +83,6 @@ namespace Data
             //modelBuilder.Properties<DateTime>().Configure(c => c.HasColumnType("datetime2"));
             modelBuilder.ApplyConfiguration(new UserConfiguration());
             modelBuilder.ApplyConfiguration(new ProjectConfiguration());
-
-            // Database seeding for demo
-            modelBuilder.SeedCollaborators(modelBuilder.SeedProjects(modelBuilder.SeedUsers()));
         }
 
     }