+++ Models/Defaults/Defaults.cs
@@ -30,6 +30,10 @@ namespace Models.Defaults
             public const string Student = "Student";
             public const string Teacher = "Teacher";
 
+            public const string RegisteredUser = "RegisteredUser";
+            public const string PrUser = "PrUser";
+            public const string Administrator = "Administrator";
+
         }
 
         public static class Claims { }
@@ -66,6 +70,17 @@ namespace Models.Defaults
             [Description("This scope gives write access to the highlight namespace")]
             public const string HighlightWrite = "highlight:write";
 
+            [Description("This scope gives read access to the Role namespace")]
+            public const string RoleRead = "role:read";
+
+            [Description("This scope gives write access to the Role namespace")]
+            public const string RoleWrite = "role:write";
+
+            [Description("This scope gives write access to the embed namespace")]
+            public const string EmbedWrite = "embed:write";
+
+            [Description("This scope gives write access to the embed namespace")]
+            public const string EmbedRead = "embed:read";
         }
 
     }